added value checks for stack cells

+++ src/stack.h
@@ -7,14 +7,14 @@
 
 
 typedef struct {
-    int8_t* arr;
-    int top;
+    int16_t* arr;
+    uint8_t top;
 } Stack;
 
 void init(Stack *stack, int size);
 int8_t is_empty(Stack *stack);
 int8_t is_full(Stack *stack);
-void push(char **error, Stack *stack, int8_t value);
+void push(char **error, Stack *stack, int16_t value);
 int8_t pop(char **error, Stack *stack);
 int8_t peek(char **error, Stack *stack);
